requirements.txt.txt
tk
mysql-connector-python
pandas
matplotlib
numpy